requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
attrs==15.2.0
Babel==2.2.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
billiard==3.3.0.23
boto==2.39.0
bz2file==0.98
celery==3.1.23
celerymon==1.0.3
certifi==2016.2.28
cffi==1.5.2
CherryPy==5.0.0
configparser==3.3.0.post2
contextlib2==0.5.1
cryptography==1.2.3
cssselect==0.9.1
cycler==0.10.0
Cython==0.24
decorator==4.0.9
Django==1.9.5
django-celery==3.1.17
django-extensions==1.6.1
django-grappelli==2.8.1
-e git://github.com/django-haystack/django-haystack.git@6e559dedbc8f17a80d361a05e16e8ce08bb09662#egg=django_haystack
django-matrix-field==0.2
django-queryset-iterator==0.1.1
django-redis-cache==1.6.5
django-webpack-loader==0.3.0
djangorestframework==3.3.3
DjangoRestless==0.0.10
docutils==0.12
elasticsearch==2.3.0
elasticstack==0.2.0
entrypoints==0.2
enum34==1.1.2
feedfinder2==0.0.1
feedparser==5.1.3
Flask==0.10.1
flower==0.9.0
funcsigs==0.4
functools32==3.2.3.post2
futures==3.0.5
gensim==0.12.4
gnp==0.0.4
gnureadline==6.3.3
Google-Search-API==1.1.11
googlemaps==2.4.3
gspread==0.3.0
gunicorn==19.4.5
httplib2==0.9.2
httpretty==0.8.10
idna==2.0
ipaddress==1.0.16
ipykernel==4.3.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
jieba==0.35
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
kombu==3.0.35
langdetect==1.0.5
line-profiler==1.0
lxml==3.3.5
MarkupSafe==0.23
matplotlib==1.5.1
meld3==1.0.2
mistune==0.7.2
mock==1.3.0
MySQL-python==1.2.5
nbconvert==4.2.0
nbformat==4.0.1
newspaper==0.0.9.8
nltk==3.2
nose==1.3.7
notebook==4.2.0
numpy==1.10.4
oauth2client==2.1.0
oauthlib==1.0.3
optional-django==0.3.0
pandas==0.18.0
Paste==2.0.2
pathlib2==2.1.0
patsy==0.4.1
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.7.2
-e git://github.com/python-pillow/Pillow.git@4d234fb54dd0d5813da636626850224485db7fe8#egg=Pillow
pip-review==0.4
plotly==1.9.2
progressbar==2.3
psycopg2==2.6.1
ptyprocess==0.5.1
py-web-search==0.2.4
py4j==0.9.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pydot==1.0.2
pyenchant==1.6.6
Pygments==2.1.3
pyOpenSSL==0.15.1
pyparsing==2.1.1
pytesseract==0.1.6
python-dateutil==2.5.1
python-ptrace==0.8.1
python-twitter==2.2
pytz==2016.4
PyYAML==3.11
pyzmq==15.2.0
qtconsole==4.2.1
queuelib==1.4.2
raven==5.12.0
react==3.0.1
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
retrying==1.3.3
rsa==3.4.2
scikit-learn==0.17.1
scipy==0.17.0
Scrapy==1.0.5
selenium==2.53.1
service-identity==16.0.0
simplegeneric==0.8.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
slackweb==1.0.5
smart-open==1.3.2
statistics==1.0.3.5
statsmodels==0.6.1
supervisor==3.2.2
terminado==0.6
tldextract==1.5.1
tornado==4.3
traitlets==4.2.1
trend-scorer==1.0.0
Twisted==16.0.0
TwitterSearch==1.0.1
Unidecode==0.4.19
urllib3==1.14
w3lib==1.13.0
Werkzeug==0.11.5
wikipedia==1.4.0
xvfbwrapper==0.2.8
zope.interface==4.1.3
